pyproject.toml
[project]
name = "bee-tag-corrector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.3",
    "python-ffmpeg>=2.0.12",
    "streamlit>=1.44.0",
    "tqdm>=4.67.1",
]
